requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.45.0
accelerate>=0.26.0
Pillow>=10.0.0
pyyaml>=6.0
einops>=0.7.0
